+++ mainchain/api.go
@@ -29,6 +29,7 @@ import (
 	"github.com/kardiachain/go-kardia/kvm"
 	"github.com/kardiachain/go-kardia/lib/abi"
 	"github.com/kardiachain/go-kardia/lib/common"
+	"github.com/kardiachain/go-kardia/lib/crypto"
 	"github.com/kardiachain/go-kardia/lib/log"
 	"github.com/kardiachain/go-kardia/lib/rlp"
 	"github.com/kardiachain/go-kardia/mainchain/blockchain"
@@ -478,6 +479,82 @@ func (s *PublicKaiAPI) KardiaCall(ctx context.Context, args types.CallArgsJSON, 
 	return result.Return(), result.Err
 }
 
+// GetValidatorSet get the validators set at block height
+func (s *PublicKaiAPI) GetValidatorSet(blockHeight rpc.BlockHeight) (*types.ValidatorSet, error) {
+	return s.kaiService.stateDB.LoadValidators(blockHeight.Uint64())
+}
+
+// GetCommit get validators' commits for the block by height
+func (s *PublicKaiAPI) GetCommit(blockHeight rpc.BlockHeight) *types.Commit {
+	return s.kaiService.kaiDb.ReadCommit(blockHeight.Uint64())
+}
+
+// Result structs for GetProof
+type AccountResult struct {
+	Address      common.Address  `json:"address"`
+	AccountProof []string        `json:"accountProof"`
+	Balance      *big.Int        `json:"balance"`
+	CodeHash     common.Hash     `json:"codeHash"`
+	Nonce        uint64          `json:"nonce"`
+	StorageHash  common.Hash     `json:"storageHash"`
+	StorageProof []StorageResult `json:"storageProof"`
+}
+type StorageResult struct {
+	Key   string   `json:"key"`
+	Value *big.Int `json:"value"`
+	Proof []string `json:"proof"`
+}
+
+// GetProof returns the Merkle-proof for a given account and optionally some storage keys.
+func (s *PublicKaiAPI) GetProof(ctx context.Context, address common.Address, storageKeys []string, blockHeightOrHash rpc.BlockHeightOrHash) (*AccountResult, error) {
+	state, _, err := s.kaiService.StateAndHeaderByHeightOrHash(ctx, blockHeightOrHash)
+	if state == nil || err != nil {
+		return nil, err
+	}
+
+	storageTrie := state.StorageTrie(address)
+	storageHash := types.EmptyRootHash
+	codeHash := state.GetCodeHash(address)
+	storageProof := make([]StorageResult, len(storageKeys))
+
+	// if we have a storageTrie, (which means the account exists), we can update the storageHash
+	if storageTrie != nil {
+		storageHash = storageTrie.Hash()
+	} else {
+		// no storageTrie means the account does not exist, so the codeHash is the hash of an empty bytearray.
+		codeHash = crypto.Keccak256Hash(nil)
+	}
+
+	// create the proof for the storageKeys
+	for i, key := range storageKeys {
+		if storageTrie != nil {
+			proof, storageError := state.GetStorageProof(address, common.HexToHash(key))
+			if storageError != nil {
+				return nil, storageError
+			}
+			storageProof[i] = StorageResult{key, state.GetState(address, common.HexToHash(key)).Big(), toHexSlice(proof)}
+		} else {
+			storageProof[i] = StorageResult{key, new(big.Int), []string{}}
+		}
+	}
+
+	// create the accountProof
+	accountProof, proofErr := state.GetProof(address)
+	if proofErr != nil {
+		return nil, proofErr
+	}
+
+	return &AccountResult{
+		Address:      address,
+		AccountProof: toHexSlice(accountProof),
+		Balance:      state.GetBalance(address),
+		CodeHash:     codeHash,
+		Nonce:        state.GetNonce(address),
+		StorageHash:  storageHash,
+		StorageProof: storageProof,
+	}, state.Error()
+}
+
 // PendingTransactions returns pending transactions
 func (a *PublicTransactionAPI) PendingTransactions() ([]*PublicTransaction, error) {
 	pendingTxs := a.s.TxPool().GetPendingData()
@@ -505,21 +582,6 @@ func (a *PublicTransactionAPI) GetTransaction(hash string) (*PublicTransaction, 
 	// get block time from block
 	publicTx.Time = block.Header().Time
 	return publicTx, nil
-}
-
-// GetTransactionCount the number of transactions the given address has sent for the given block number
-func (a *PublicTransactionAPI) GetTransactionCount(ctx context.Context, address common.Address, blockHeightOrHash rpc.BlockHeightOrHash) (uint64, error) {
-	// Ask transaction pool for the nonce which includes pending transactions
-	if blockHeight, ok := blockHeightOrHash.Height(); ok && blockHeight == rpc.PendingBlockHeight {
-		return a.s.txPool.Nonce(address), nil
-	}
-	// Resolve block number and use its state to ask for the nonce
-	state, _, err := a.s.StateAndHeaderByHeightOrHash(ctx, blockHeightOrHash)
-	if state == nil || err != nil {
-		return 0, err
-	}
-	nonce := state.GetNonce(address)
-	return nonce, state.Error()
 }
 
 // getReceiptLogs gets logs from receipt
@@ -634,6 +696,21 @@ func (a *PublicAccountAPI) Nonce(address string) (uint64, error) {
 	addr := common.HexToAddress(address)
 	nonce := a.kaiService.txPool.Nonce(addr)
 	return nonce, nil
+}
+
+// NonceAtHeight the nonce of an address at the given height
+func (a *PublicAccountAPI) NonceAtHeight(ctx context.Context, address common.Address, blockHeightOrHash rpc.BlockHeightOrHash) (uint64, error) {
+	// Ask transaction pool for the nonce, which includes pending transactions
+	if blockHeight, ok := blockHeightOrHash.Height(); ok && blockHeight == rpc.PendingBlockHeight {
+		return a.kaiService.txPool.Nonce(address), nil
+	}
+	// Resolve block height and use its state to ask for the nonce
+	state, _, err := a.kaiService.StateAndHeaderByHeightOrHash(ctx, blockHeightOrHash)
+	if state == nil || err != nil {
+		return 0, err
+	}
+	nonce := state.GetNonce(address)
+	return nonce, state.Error()
 }
 
 // GetCode returns the code stored at the given address in the state for the given block height.
@@ -820,4 +897,13 @@ func checkGas(gasPrice *big.Int, gas uint64) error {
 		return ErrExceedGasLimit
 	}
 	return nil
+}
+
+// toHexSlice creates a slice of hex-strings based on []byte.
+func toHexSlice(b [][]byte) []string {
+	r := make([]string, len(b))
+	for i := range b {
+		r[i] = common.Encode(b[i])
+	}
+	return r
 }
