--- mainchain/verify_api.go
@@ -1,121 +1,0 @@
-/*
- *  Copyright 2021 KardiaChain
- *  This file is part of the go-kardia library.
- *
- *  The go-kardia library is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU Lesser General Public License as published by
- *  the Free Software Foundation, either version 3 of the License, or
- *  (at your option) any later version.
- *
- *  The go-kardia library is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- *  GNU Lesser General Public License for more details.
- *
- *  You should have received a copy of the GNU Lesser General Public License
- *  along with the go-kardia library. If not, see <http://www.gnu.org/licenses/>.
- */
-
-package kai
-
-import (
-	"context"
-	"math/big"
-
-	"github.com/kardiachain/go-kardia/lib/common"
-	"github.com/kardiachain/go-kardia/lib/crypto"
-	"github.com/kardiachain/go-kardia/rpc"
-	"github.com/kardiachain/go-kardia/types"
-)
-
-// VerifyKaiAPI provides APIs to access Kai full node-related information.
-type VerifyKaiAPI struct {
-	k *KardiaService
-}
-
-// NewVerifyKaiAPI creates a new Kai protocol API for full nodes.
-func NewVerifyKaiAPI(kaiService *KardiaService) *VerifyKaiAPI {
-	return &VerifyKaiAPI{kaiService}
-}
-
-func (v *VerifyKaiAPI) GetValidatorSet(blockHeight rpc.BlockHeight) (*types.ValidatorSet, error) {
-	return v.k.stateDB.LoadValidators(blockHeight.Uint64())
-}
-
-func (v *VerifyKaiAPI) GetCommit(blockHeight rpc.BlockHeight) *types.Commit {
-	return v.k.kaiDb.ReadCommit(blockHeight.Uint64())
-}
-
-// Result structs for GetProof
-type AccountResult struct {
-	Address      common.Address  `json:"address"`
-	AccountProof []string        `json:"accountProof"`
-	Balance      *big.Int        `json:"balance"`
-	CodeHash     common.Hash     `json:"codeHash"`
-	Nonce        uint64          `json:"nonce"`
-	StorageHash  common.Hash     `json:"storageHash"`
-	StorageProof []StorageResult `json:"storageProof"`
-}
-type StorageResult struct {
-	Key   string   `json:"key"`
-	Value *big.Int `json:"value"`
-	Proof []string `json:"proof"`
-}
-
-func (v *VerifyKaiAPI) GetProof(ctx context.Context, address common.Address, storageKeys []string, blockHeightOrHash rpc.BlockHeightOrHash) (*AccountResult, error) {
-	state, _, err := v.k.StateAndHeaderByHeightOrHash(ctx, blockHeightOrHash)
-	if state == nil || err != nil {
-		return nil, err
-	}
-
-	storageTrie := state.StorageTrie(address)
-	storageHash := types.EmptyRootHash
-	codeHash := state.GetCodeHash(address)
-	storageProof := make([]StorageResult, len(storageKeys))
-
-	// if we have a storageTrie, (which means the account exists), we can update the storageHash
-	if storageTrie != nil {
-		storageHash = storageTrie.Hash()
-	} else {
-		// no storageTrie means the account does not exist, so the codeHash is the hash of an empty bytearray.
-		codeHash = crypto.Keccak256Hash(nil)
-	}
-
-	// create the proof for the storageKeys
-	for i, key := range storageKeys {
-		if storageTrie != nil {
-			proof, storageError := state.GetStorageProof(address, common.HexToHash(key))
-			if storageError != nil {
-				return nil, storageError
-			}
-			storageProof[i] = StorageResult{key, state.GetState(address, common.HexToHash(key)).Big(), toHexSlice(proof)}
-		} else {
-			storageProof[i] = StorageResult{key, new(big.Int), []string{}}
-		}
-	}
-
-	// create the accountProof
-	accountProof, proofErr := state.GetProof(address)
-	if proofErr != nil {
-		return nil, proofErr
-	}
-
-	return &AccountResult{
-		Address:      address,
-		AccountProof: toHexSlice(accountProof),
-		Balance:      state.GetBalance(address),
-		CodeHash:     codeHash,
-		Nonce:        state.GetNonce(address),
-		StorageHash:  storageHash,
-		StorageProof: storageProof,
-	}, state.Error()
-}
-
-// toHexSlice creates a slice of hex-strings based on []byte.
-func toHexSlice(b [][]byte) []string {
-	r := make([]string, len(b))
-	for i := range b {
-		r[i] = common.Encode(b[i])
-	}
-	return r
-}
